+++ frontend/src/components/google-maps/map/CarFfeineListener.tsx
@@ -1,0 +1,58 @@
+import { useEffect } from 'react';
+
+import { useQueryClient } from '@tanstack/react-query';
+
+import { debounce } from '@utils/debounce';
+import { useExternalValue, useSetExternalState } from '@utils/external-state';
+import { getDisplayPosition } from '@utils/google-maps';
+import { isCachedRegion } from '@utils/google-maps/isCachedRegion';
+import { setLocalStorage } from '@utils/storage';
+
+import { getGoogleMapStore } from '@stores/google-maps/googleMapStore';
+import { warningModalActions } from '@stores/layout/warningModalStore';
+import { popupMenuOpenStore } from '@stores/popupMenuOpenStore';
+
+import ZoomWarningModal from '@common/WarningModal';
+
+import { INITIAL_ZOOM_SIZE } from '@constants/googleMaps';
+import { QUERY_KEY_STATION_MARKERS } from '@constants/queryKeys';
+import { LOCAL_KEY_LAST_POSITION } from '@constants/storageKeys';
+
+const CarFfeineMapListener = () => {
+  const googleMap = useExternalValue(getGoogleMapStore());
+  const queryClient = useQueryClient();
+  const setIsPopupMenuOpen = useSetExternalState(popupMenuOpenStore);
+
+  const debouncedIdleHandler = debounce(() => {
+    if (googleMap.getZoom() < INITIAL_ZOOM_SIZE) {
+      warningModalActions.openModal(<ZoomWarningModal />);
+    } else {
+      warningModalActions.closeModal();
+    }
+    const displayPosition = getDisplayPosition(googleMap);
+    if (!isCachedRegion(displayPosition)) {
+      queryClient.invalidateQueries({ queryKey: [QUERY_KEY_STATION_MARKERS] });
+    }
+
+    setIsPopupMenuOpen(false);
+
+    setLocalStorage<google.maps.LatLngLiteral>(LOCAL_KEY_LAST_POSITION, {
+      lat: googleMap.getCenter().lat(),
+      lng: googleMap.getCenter().lng(),
+    });
+  }, 300);
+
+  useEffect(() => {
+    googleMap.addListener('idle', debouncedIdleHandler);
+
+    const initMarkersEvent = googleMap.addListener('bounds_changed', async () => {
+      queryClient.invalidateQueries({ queryKey: [QUERY_KEY_STATION_MARKERS] });
+
+      google.maps.event.removeListener(initMarkersEvent);
+    });
+  }, []);
+
+  return <></>;
+};
+
+export default CarFfeineMapListener;
